refactor: PresignedUrl 생성 메서드 유틸클래스로 이동
