+++ nmnb-application/src/main/kotlin/nmnb/application/global/infrastructure/external/s3/S3Service.kt
@@ -11,14 +11,10 @@ import software.amazon.awssdk.core.sync.RequestBody
 import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.GetObjectRequest
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
-import software.amazon.awssdk.services.s3.presigner.S3Presigner
-import software.amazon.awssdk.services.s3.presigner.model.GetObjectPresignRequest
-import java.time.Duration
 
 @Service
 class S3Service(
     private val s3Client: S3Client,
-    private val s3Presigner: S3Presigner,
     private val s3Properties: S3Properties,
     private val s3Utils: S3Utils,
 ) {
@@ -32,7 +28,7 @@ class S3Service(
             throw S3Exception(ErrorStatus.S3_UPLOAD_PROFILE_IMAGE_FAILED)
         }
 
-        return generatePresignedUrl(s3Key)
+        return s3Utils.generatePresignedUrl(s3Key)
     }
 
     private fun createRequest(s3Key: String, profileImage: MultipartFile): PutObjectRequest =
@@ -43,22 +39,7 @@ class S3Service(
             .contentLength(profileImage.size)
             .build()
 
-    private fun generatePresignedUrl(key: String): String {
-        val getObjectRequest = GetObjectRequest.builder()
-            .bucket(s3Properties.s3.bucket)
-            .key(key)
-            .build()
-
-        val presignRequest = GetObjectPresignRequest.builder()
-            .getObjectRequest(getObjectRequest)
-            .signatureDuration(PRESIGNED_URL_EXPIRATION)
-            .build()
-
-        return s3Presigner.presignGetObject(presignRequest).url().toString()
-    }
-
     companion object {
-        private val PRESIGNED_URL_EXPIRATION: Duration = Duration.ofHours(1)
         private const val PROFILE_IMAGE_PATH = "profile"
     }
 }
